+++ src-inst/flow_storm/api.cljs
@@ -43,8 +43,9 @@
   ([dw-id val stack-key extra]
    (let [vdata (assoc (rt-values/extract-data-aspects val extra)
                       :flow-storm.debugger.ui.data-windows.data-windows/dw-id dw-id
-                      :flow-storm.debugger.ui.data-windows.data-windows/stack-key stack-key)]
-     (rt-events/publish-event! (rt-events/make-data-window-push-val-data-event dw-id vdata true)))))
+                      :flow-storm.debugger.ui.data-windows.data-windows/stack-key stack-key)
+         extra (update extra :root? (fn [root?] (if (nil? root?) true root?)))]
+     (rt-events/publish-event! (rt-events/make-data-window-push-val-data-event dw-id vdata extra)))))
 
 (defn data-window-val-update [dw-id new-val]
   (rt-events/publish-event! (rt-events/make-data-window-update-event dw-id {:new-val new-val})))