+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
     - Add configurable auto jump to exceptions
     - New eql-query-pprint visualizer
+    - Add support for setting the visualizer via flow-storm.api/data-window-push-val
     
 ### Changes